Add files via upload
Fixed issue:
When choosing the exact day on which DST time switches to Standard time or vice versa, the calculator uses the old DST status, because timestamp 00:00 h is used. The time is switched at 1 am or 2am usually, but not at 0am. So the compare date to find out the DST status gets the timestamp 12:00.