Add files via upload
This program shows instant sun data of your current location at the current time, no matter where in the world you are.

The program works offline, you only need GPS enabled. A correct timezone recognition is implemented, too.

This program was originally supposed to be a widget but Pythonista won‘t display it because it is too complex.

Instead, I released it as a regular Pythonista Script.